pyproject.toml
[project]
name = "andrew"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.9.2",
    "openai>=1.52.0",
    "python-dotenv>=1.0.1",
    "rich>=13.9.2",
    "ruff>=0.7.0",
    "seaborn>=0.13.2",
]
